cleaned up scripts
